+++ Libs/CRC8/CRC8.h
@@ -4,6 +4,6 @@
 
 #include "stdint.h"
 
-uint8_t crc8(const uint8_t *addr, uint8_t len, uint8_t crc_init); 
+uint8_t crc8(const uint8_t *addr, uint16_t len, uint8_t crc_init); 
 
 #endif
